addressed comments from @bilderbuchi
