+++ pymeasure/instruments/__init__.py
@@ -24,10 +24,9 @@
 
 from ..errors import RangeError, RangeException
 from .instrument import Instrument
-from .instrument import FakeInstrument
-from .mock import Mock
 from .resources import list_resources
 from .validators import discreteTruncate
+from .fakes import SwissArmyFake, FakeInstrument
 
 from . import advantest
 from . import agilent
