+++ pymeasure/instruments/fakes.py
@@ -22,17 +22,67 @@
 # THE SOFTWARE.
 #
 
+import re
 import time
 import numpy as np
 
 from pymeasure.adapters import FakeAdapter
-from pymeasure.instruments import FakeInstrument
+from pymeasure.instruments import Instrument
 from pymeasure.instruments.validators import strict_discrete_set
 
 
-class Mock(FakeInstrument):
-    """Mock instrument useful for testing. Mock data that can be generated by this class includes voltages, waveforms,
-       and image data.
+class FakeInstrument(Instrument):
+    """ Provides a fake implementation of the Instrument class
+    for testing purposes.
+    """
+
+    def __init__(self, adapter=None, name=None, includeSCPI=False, **kwargs):
+        super().__init__(
+            FakeAdapter(**kwargs),
+            name or "Fake Instrument",
+            includeSCPI=includeSCPI,
+            **kwargs
+        )
+
+    @staticmethod
+    def control(get_command, set_command, docs,
+                validator=lambda v, vs: v, values=(), map_values=False,
+                get_process=lambda v: v, set_process=lambda v: v,
+                check_set_errors=False, check_get_errors=False,
+                **kwargs):
+        """Fake Instrument.control.
+
+        Strip commands and only store and return values indicated by
+        format strings to mimic many simple commands.
+        This is analogous how the tests in test_instrument are handled.
+        """
+
+        # Regex search to find first format specifier in the command
+        fmt_spec_pattern = r'(%[\w.#-+ *]*[diouxXeEfFgGcrsa%])'
+        match = re.findall(fmt_spec_pattern, set_command)
+        if match:
+            # format_specifier = match.group(0)
+            format_specifier = ','.join(match)
+        else:
+            format_specifier = ''
+        # To preserve as much functionality as possible, call the real
+        # control method with modified get_command and set_command.
+        return Instrument.control(get_command="",
+                                  set_command=format_specifier,
+                                  docs=docs,
+                                  validator=validator,
+                                  values=values,
+                                  map_values=map_values,
+                                  get_process=get_process,
+                                  set_process=set_process,
+                                  check_set_errors=check_set_errors,
+                                  check_get_errors=check_get_errors,
+                                  **kwargs)
+
+
+class SwissArmyFake(FakeInstrument):
+    """Fake instrument useful for testing. Mock data that can currently be generated by this class includes voltages,
+       waveforms, and image data.
     """
 
     def __init__(self, wait=.1, **kwargs):
@@ -138,7 +188,8 @@ class Mock(FakeInstrument):
         im_format_type_dict = {"8": np.uint8, "16": np.uint16} 
         bit_depth = self.frame_format.split("_")[1]
         time.sleep(self._wait)
-        return np.array(im_format_maxval_dict[bit_depth] * np.random.rand(self.frame_height, self.frame_width), dtype=im_format_type_dict[bit_depth]) 
+        return np.array(im_format_maxval_dict[bit_depth] * np.random.rand(self.frame_height, self.frame_width),
+                        dtype=im_format_type_dict[bit_depth])
 
     def __getitem__(self, keys):
         return keys
