+++ tests/instruments/test_instrument.py
@@ -23,12 +23,13 @@
 #
 
 import pytest
-from pymeasure.instruments.instrument import Instrument, FakeInstrument
+from pymeasure.instruments.fakes import SwissArmyFake
+from pymeasure.instruments.instrument import Instrument
 from pymeasure.instruments.validators import strict_discrete_set, strict_range
 
 
 def test_fake_instrument():
-    fake = FakeInstrument()
+    fake = SwissArmyFake()
     fake.write("Testing")
     assert fake.read() == "Testing"
     assert fake.read() == ""
@@ -47,7 +48,7 @@ def test_control_doc():
 
 
 def test_control_validator():
-    class Fake(FakeInstrument):
+    class Fake(SwissArmyFake):
         x = Instrument.control(
             "", "%d", "",
             validator=strict_discrete_set,
@@ -64,7 +65,7 @@ def test_control_validator():
 
 
 def test_control_validator_map():
-    class Fake(FakeInstrument):
+    class Fake(SwissArmyFake):
         x = Instrument.control(
             "", "%d", "",
             validator=strict_discrete_set,
@@ -82,7 +83,7 @@ def test_control_validator_map():
 
 
 def test_control_dict_map():
-    class Fake(FakeInstrument):
+    class Fake(SwissArmyFake):
         x = Instrument.control(
             "", "%d", "",
             validator=strict_discrete_set,
@@ -100,7 +101,7 @@ def test_control_dict_map():
 
 
 def test_control_dict_str_map():
-    class Fake(FakeInstrument):
+    class Fake(SwissArmyFake):
         x = Instrument.control(
             "", "%d", "",
             validator=strict_discrete_set,
@@ -118,7 +119,7 @@ def test_control_dict_str_map():
 
 
 def test_control_process():
-    class Fake(FakeInstrument):
+    class Fake(SwissArmyFake):
         x = Instrument.control(
             "", "%d", "",
             validator=strict_range,
@@ -135,7 +136,7 @@ def test_control_process():
 
 
 def test_control_get_process():
-    class Fake(FakeInstrument):
+    class Fake(SwissArmyFake):
         x = Instrument.control(
             "", "JUNK%d", "",
             validator=strict_range,
@@ -152,7 +153,7 @@ def test_control_get_process():
 
 def test_control_preprocess_reply_property():
     # test setting preprocess_reply at property-level
-    class Fake(FakeInstrument):
+    class Fake(SwissArmyFake):
         x = Instrument.control(
             "", "JUNK%d",
             "",
@@ -173,7 +174,7 @@ def test_control_preprocess_reply_property():
 
 def test_control_preprocess_reply_adapter():
     # test setting preprocess_reply at Adapter-level
-    class Fake(FakeInstrument):
+    class Fake(SwissArmyFake):
         def __init__(self):
             super().__init__(preprocess_reply=lambda v: v.replace('JUNK', ''))
 
@@ -192,7 +193,7 @@ def test_control_preprocess_reply_adapter():
 
 
 def test_measurement_dict_str_map():
-    class Fake(FakeInstrument):
+    class Fake(SwissArmyFake):
         x = Instrument.measurement(
             "", "",
             values={'X': 1, 'Y': 2, 'Z': 3},
@@ -209,7 +210,7 @@ def test_measurement_dict_str_map():
 
 
 def test_setting_process():
-    class Fake(FakeInstrument):
+    class Fake(SwissArmyFake):
         x = Instrument.setting(
             "OUT %d", "",
             set_process=lambda v: int(bool(v)),
@@ -223,7 +224,7 @@ def test_setting_process():
 
 
 def test_control_multivalue():
-    class Fake(FakeInstrument):
+    class Fake(SwissArmyFake):
         x = Instrument.control(
             "", "%d,%d", "",
         )
@@ -239,10 +240,10 @@ def test_control_multivalue():
      ("%d, %d", (5, 6), [5, 6]),  # input has to be a tuple, not a list
      ])
 def test_fakeinstrument_control(set_command, given, expected):
-    """FakeInstrument's custom simple control needs to process values correctly.
+    """SwissArmyFake's custom simple control needs to process values correctly.
     """
-    class Fake(FakeInstrument):
-        x = FakeInstrument.control(
+    class Fake(SwissArmyFake):
+        x = SwissArmyFake.control(
             "", set_command, "",
         )
